feat: Added the ability to queue multiple GROMACS production (md) jobs if the job scheduler walltime is too short.
This runs through free npt, and uses free npt's performance along with md's mdp file to estimate how many job submissions are needed to finish md, using checkpointing to continue where the previous allocation left off. Additionally, several options were added for toggling automatic lambda job submission and enabling this checkpointing feature. An md job is queued at the same time as the first "setup jobs" with dependencies to allow it to age in the queue.

BREAKING CHANGE: The checkpointing/job submission feature is enabled by default, as this is likely to be used more often than not.